Portability - compilation warning on Mac OS/X removed
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@8138 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ manager.c
@@ -1533,7 +1533,7 @@ int manager_event(int category, const char *event, const char *fmt, ...)
 			if (timestampevents) {
 				now = ast_tvnow();
 				ast_build_string(&tmp_next, &tmp_left, "Timestamp: %ld.%06lu\r\n",
-						 now.tv_sec, now.tv_usec);
+						 now.tv_sec, (unsigned long) now.tv_usec);
 			}
 			va_start(ap, fmt);
 			ast_build_string_va(&tmp_next, &tmp_left, fmt, ap);
